% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sd.R
\name{extract_sd}
\alias{extract_sd}
\title{Extract sd}
\usage{
extract_sd(nc_path, vars)
}
\arguments{
\item{nc_path}{the path to your file}

\item{vars}{a vector containing the BGC variables you want in the output table}
}
\value{
a table with your BGC variables, depth, date, lon and lat
}
\description{
Create a dataframe with bgc variables values, depth, date, lon and lat from a SD profile.
}
