% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date.R
\name{get_date}
\alias{get_date}
\title{Date of profile}
\usage{
get_date(nc)
}
\arguments{
\item{nc}{is a nc object returned by the nc_open of ncdf4 package}
}
\value{
tge function return a date with YYYY/MM/DD format
}
\description{
A function to extract the date of the profile
}
