% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qc.R
\name{get_qc}
\alias{get_qc}
\title{Extract QC of variables}
\usage{
get_qc(vars, nc_path)
}
\arguments{
\item{vars}{is a vector of variable}

\item{nc_path}{is the path of the ncdf file}
}
\value{
a tbale with QC value of each variable along the depth
}
\description{
Create a data frame with the QC of selected variable along the depth of the profile
}
